--- facet-reflect/src/poke/mod.rs
@@ -1,37 +1,0 @@
-extern crate alloc;
-
-mod builder;
-pub use builder::*;
-
-mod iset;
-pub use iset::*;
-
-mod value_uninit;
-pub use value_uninit::*;
-
-mod value;
-pub use value::*;
-
-mod list_uninit;
-pub use list_uninit::*;
-
-mod list;
-pub use list::*;
-
-mod map_uninit;
-pub use map_uninit::*;
-
-mod map;
-pub use map::*;
-
-mod option_uninit;
-pub use option_uninit::*;
-
-mod option;
-pub use option::*;
-
-mod smart_pointer_uninit;
-pub use smart_pointer_uninit::*;
-
-mod smart_pointer;
-pub use smart_pointer::*;
