+++ facet-reflect/tests/peek/mod.rs
@@ -1,4 +1,6 @@
+mod c_test;
 mod enum_;
+mod facts;
 mod list;
 mod map;
 mod option;
