+++ facet-reflect/tests/wip/mod.rs
@@ -1,5 +1,4 @@
 mod builder;
 mod c_test;
-mod facts;
 mod ui_tests;
 mod value;
